+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 3.0.21
  - Convert ID to Names fixes
+ - new "Add Missing Parentheses to the Active Editor" and "Add Missing Parentheses to the Active Project" commands
+ - NullReference error in "Remove Unused Variables" commands
 
 ## 3.0.20
  - Issue #290 - Add application areas and captions to page labels